fix(tests): Fixed wrong tests. Correct serialized byte lengths in network and operation modes modules tests

+++ src/main/java/com/acousea/backend/core/communicationSystem/domain/nodes/extModules/operationModes/OperationModesModule.java
@@ -94,7 +94,7 @@ public class OperationModesModule extends SerializableModule implements ExtModul
     }
 
     public void removeOperationMode(byte id) {
-        int unsignedId = Byte.toUnsignedInt(id);
+        short unsignedId = (short) Byte.toUnsignedInt(id);
         if (!modes.containsKey(unsignedId)) {
             throw new IllegalArgumentException("Operation mode ID not found: " + id);
         }
